backend/requirements.txt
# Use Python 3.11–3.13 recommended; 3.14 needs recent pydantic with prebuilt wheels.
fastapi>=0.115.0,<0.116
uvicorn[standard]>=0.32.0,<0.35
sqlalchemy>=2.0.30,<3
psycopg[binary]>=3.2.13,<4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<4.1
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic[email]>=2.12.0,<3
alembic>=1.13.0,<2
email-validator>=2.1.0
requests>=2.31.0
apscheduler>=3.10.0,<4
